feat(test-observability): added logs to connect memory database (#116)
* feat(test-observability): added logs to connect memory database

* feat(test-observability): async await

* feat(test-observability): avoid call dropDatabase and clearDatabase if not running mongod

* feat(test-observability): removing async connect call

* feat(test-observability): timeout 20000ms

+++ test/database/memoryDatabase.ts
@@ -8,11 +8,12 @@ export const databaseTestModule = () => {
    * Connect to the in-memory database.
    */
   const connect = async () => {
+    console.log('connect');
     await mongod.start();
 
     // https://github.com/nodkz/mongodb-memory-server/commit/9abf04f23188e1ad4eb47b0797c33e8210b8056b#diff-45cc1c887ca9f666fee44d0273bab3652241b6bb8f8908fca5265264f3203e68R264
     // use getUri instead of deprecated getConnectionString
-    const uri = await mongod.getUri();
+    const uri = mongod.getUri();
 
     const mongooseOpts = {
       useNewUrlParser: true,
@@ -29,6 +30,14 @@ export const databaseTestModule = () => {
    * Drop database, close the connection and stop mongod.
    */
   const closeDatabase = async () => {
+    const mongodState = mongod.state;
+    if (mongodState !== 'running') {
+      console.log({
+        mongodState,
+      });
+      return;
+    }
+
     await mongoose.connection.dropDatabase();
     await mongoose.connection.close();
     await mongod.stop();
@@ -38,6 +47,14 @@ export const databaseTestModule = () => {
    * Remove all the data for all db collections.
    */
   const clearDatabase = async () => {
+    const mongodState = mongod.state;
+    if (mongodState !== 'running') {
+      console.log({
+        mongodState,
+      });
+      return;
+    }
+
     const collections = mongoose.connection.collections;
 
     for (const key in collections) {
@@ -52,5 +69,6 @@ export const databaseTestModule = () => {
     connect,
     closeDatabase,
     clearDatabase,
+    mongod,
   };
 };
